restructured the files for better organization
