+++ src/components/LoadingScreen.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+/*
+ * HTML to default to when fridge data cannot be retrieved.
+ * return - a <div> that tells the user the page is loading
+ */
+const LoadingScreen = () => {
+    return <div>Loading</div>
+}
+
+export default LoadingScreen
