+++ src/components/MapControls.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+/*
+ * Buttons layered over the map to give the user control of the data view.
+ * return - the button HTML to display on a <Map />
+ */
+const MapControls  = () => {
+    return <div>Map Controls</div>
+}
+
+export default MapControls
