--- src/fridgePopupInfo.js
@@ -1,6 +1,0 @@
-
-export default function fridgePopupInfo() {
-    return (
-        <div></div>
-    )
-}
