+++ src/images/mapLocationIcon.png
