+++ src/images/mapLocationIconBlack.png
